Update backend crate names
