+++ backends/js_wasm_runtime_layer/README.md
@@ -1,0 +1,6 @@
+# js_wasm_runtime_layer
+
+[![Crates.io](https://img.shields.io/crates/v/js_wasm_runtime_layer.svg)](https://crates.io/crates/js_wasm_runtime_layer)
+[![Docs.rs](https://docs.rs/js_wasm_runtime_layer/badge.svg)](https://docs.rs/js_wasm_runtime_layer)
+
+`js_wasm_runtime_layer` implements the `wasm_runtime_layer` abstraction interface over WebAssembly runtimes for your web browser's WebAssembly runtime.
