+++ backends/js_wasm_runtime_layer/src/table.rs
