--- backends/wasmi-runtime-layer/README.md
@@ -1,6 +1,0 @@
-# wasmi-runtime-layer
-
-[![Crates.io](https://img.shields.io/crates/v/wasmi-runtime-layer.svg)](https://crates.io/crates/wasmi-runtime-layer)
-[![Docs.rs](https://docs.rs/wasmi-runtime-layer/badge.svg)](https://docs.rs/wasmi-runtime-layer)
-
-`wasmi-runtime-layer` implements the `wasm_runtime_layer` abstraction interface over WebAssembly runtimes for Wasmi.
