+++ backends/wasmi_runtime_layer/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wasmi-runtime-layer"
+name = "wasmi_runtime_layer"
 version = "0.31.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
