+++ backends/wasmi_runtime_layer/README.md
@@ -1,0 +1,6 @@
+# wasmi_runtime_layer
+
+[![Crates.io](https://img.shields.io/crates/v/wasmi_runtime_layer.svg)](https://crates.io/crates/wasmi_runtime_layer)
+[![Docs.rs](https://docs.rs/wasmi_runtime_layer/badge.svg)](https://docs.rs/wasmi_runtime_layer)
+
+`wasmi_runtime_layer` implements the `wasm_runtime_layer` abstraction interface over WebAssembly runtimes for Wasmi.
