--- backends/wasmtime-runtime-layer/README.md
@@ -1,6 +1,0 @@
-# wasmtime-runtime-layer
-
-[![Crates.io](https://img.shields.io/crates/v/wasmtime-runtime-layer.svg)](https://crates.io/crates/wasmtime-runtime-layer)
-[![Docs.rs](https://docs.rs/wasmtime-runtime-layer/badge.svg)](https://docs.rs/wasmtime-runtime-layer)
-
-`wasmtime-runtime-layer` implements the `wasm_runtime_layer` abstraction interface over WebAssembly runtimes for Wasmtime.
