+++ backends/wasmtime_runtime_layer/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wasmtime-runtime-layer"
+name = "wasmtime_runtime_layer"
 version = "18.0.0"
 edition = "2021"
 license = "MIT OR Apache-2.0"
