--- backends/webassembly-runtime-layer/README.md
@@ -1,6 +1,0 @@
-# webassembly-runtime-layer
-
-[![Crates.io](https://img.shields.io/crates/v/webassembly-runtime-layer.svg)](https://crates.io/crates/webassembly-runtime-layer)
-[![Docs.rs](https://docs.rs/webassembly-runtime-layer/badge.svg)](https://docs.rs/webassembly-runtime-layer)
-
-`webassembly-runtime-layer` implements the `wasm_runtime_layer` abstraction interface over WebAssembly runtimes for your web browser's WebAssembly runtime.
